+++ src/main/java/com/raoulvdberge/refinedstorage/tile/TileInterface.java
@@ -4,6 +4,7 @@ import com.raoulvdberge.refinedstorage.RS;
 import com.raoulvdberge.refinedstorage.RSUtils;
 import com.raoulvdberge.refinedstorage.api.util.IComparer;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerBasic;
+import com.raoulvdberge.refinedstorage.inventory.ItemHandlerInterface;
 import com.raoulvdberge.refinedstorage.inventory.ItemHandlerUpgrade;
 import com.raoulvdberge.refinedstorage.item.ItemUpgrade;
 import com.raoulvdberge.refinedstorage.tile.config.IComparable;
@@ -16,7 +17,6 @@ import net.minecraftforge.items.CapabilityItemHandler;
 import net.minecraftforge.items.IItemHandler;
 import net.minecraftforge.items.wrapper.CombinedInvWrapper;
 
-import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 public class TileInterface extends TileNode implements IComparable {
@@ -27,13 +27,9 @@ public class TileInterface extends TileNode implements IComparable {
     private ItemHandlerBasic importItems = new ItemHandlerBasic(9, this);
 
     private ItemHandlerBasic exportSpecimenItems = new ItemHandlerBasic(9, this);
-    private ItemHandlerBasic exportItems = new ItemHandlerBasic(9, this) {
-        @Override
-        @Nonnull
-        public ItemStack insertItem(int slot, @Nonnull ItemStack stack, boolean simulate) {
-            return stack;
-        }
-    };
+    private ItemHandlerBasic exportItems = new ItemHandlerBasic(9, this);
+
+    private ItemHandlerInterface items = new ItemHandlerInterface(importItems, exportItems);
 
     private ItemHandlerUpgrade upgrades = new ItemHandlerUpgrade(4, this, ItemUpgrade.TYPE_SPEED, ItemUpgrade.TYPE_STACK, ItemUpgrade.TYPE_CRAFTING);
 
@@ -193,7 +189,7 @@ public class TileInterface extends TileNode implements IComparable {
     @Override
     public <T> T getCapability(Capability<T> capability, @Nullable EnumFacing facing) {
         if (capability == CapabilityItemHandler.ITEM_HANDLER_CAPABILITY) {
-            return CapabilityItemHandler.ITEM_HANDLER_CAPABILITY.cast(facing == EnumFacing.DOWN ? exportItems : importItems);
+            return CapabilityItemHandler.ITEM_HANDLER_CAPABILITY.cast(items);
         }
 
         return super.getCapability(capability, facing);